test - Add initial test for vehicle storage
This commit adds the initial unit tests for the vehicle storage functionality.
It also refactor the welcome page code and introduces a vehicle manager class, in
order to be able to test logic without GUI.
The commit also renames the functionality directory from rcutils to rcfunc instead since it will better
support more refactoring and increase the maintainability in the future.