test - Initial unit test for maintenance functionality
This commit adds the first unit tests for the maintenance base functionality.
In order to be able to create these tests, a refactoring of the maintenance tab had to be done as well.
The refactoring splits the functionality away from the GUI parts, which makes it possible to test and maintain in
a better way.