Feat - Local bin setup script
This commit adds the local bin link script which should be used to "install" rc by the user.
It will improve accessability, and remove the need to execute it from the repo itself.

The commit also adds a new test with jobs in the commit workflow for github action in order to verify the
script execution and start of application after as a user.