Impr - Add ruff linting to project
This commit adds ruff linting to the project. It will lint the source files (rctabs, rcfunc and tests).
The commit also adds the tool directory with a wrapper script around the ruff call which can be used
to execute the linting locally. Usage instructions also provided to support the user.

A new job in github actions, commit workflow, is also added to run linting on commits pushed to the repo. This will
hopefully improve code quality. Fixes to issues which was found also added.