bugfix - Session display showing non reversed order track session data
This commit fixes a issue with the latest added track day, showing data from the
oldest stored track day. This is due to the display is done in reversed order while
the track session data is loaded with normal index, causing the data and track day sessions data to be
mismatched. The fix is done in the display function, which now uses a origin index instead of the updated one after
reversing.