Impr - Scrollbar for track day tab
This commit adds the scrolling functionality to the track day list tab, needed due to not
full list is visible when the entry amount is more then the start size of the app. In order
to be able to test the mousewheel binding, small refactoring is done also from the maintenance tab
in order to have common code. Scrolling functionality is seperated out in a new utility file for the GUI.

The commit also include a unit test for the binding.